+++ example/example_task_download_file.dart
@@ -28,7 +28,7 @@ Future<void> main() async {
   // User request to cancel
   Timer(Duration(seconds: sec), () {
     print('');
-    print('Cancelling after $sec sec');
+    print('Canceling after $sec sec');
     cts.cancel();
   });
 
@@ -56,7 +56,7 @@ Task<String> _download(Uri uri, String filename, CancellationToken token,
       _message('Downloaded: ${bytes.length} bytes');
     });
 
-    token.throwIfCancelled();
+    token.throwIfCanceled();
     final client = Client();
     final abortTrigger = Completer<void>();
 
@@ -90,7 +90,7 @@ Task<String> _download(Uri uri, String filename, CancellationToken token,
       }
     }
 
-    await token.runCancellable(abortTrigger.complete, get);
+    await token.runCancelable(abortTrigger.complete, get);
 
     // Save file to disk
     await Future<void>.delayed(Duration(seconds: 1));