+++ lib/synchronization/binary_semaphore.dart
@@ -58,8 +58,7 @@ class BinarySemaphore extends Lock {
   @useResult
   Future<bool> tryAcquire(Duration timeout) {
     if (timeout.isNegative) {
-      throw ArgumentError.value(
-          timeout, 'timeout', 'Timeout must not be negative');
+      throw ArgumentError.value(timeout, 'timeout', 'Must not be negative');
     }
 
     if (!_isLocked) {