+++ lib/synchronization/condition_variable.dart
@@ -53,8 +53,7 @@ class ConditionVariable {
   Future<bool> tryWait(Duration timeout) async {
     final started = _stopwatch.elapsedMicroseconds;
     if (timeout.isNegative) {
-      throw ArgumentError.value(
-          timeout, 'timeout', 'Timeout must not be negative');
+      throw ArgumentError.value(timeout, 'timeout', 'Must not be negative');
     }
 
     await lock.release();